backend/django/pyproject.toml
[tool.poetry]
name = "metmuseum-django-utils"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
pyjwt = "^2.9.0"
python-dotenv = "^1.0.1"
django-mysql = "^4.15.0"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
